+++ DVTFoundation/_DVTTimeSlicedMainThreadUnorderedUniquingWorkQueue.h
@@ -17,7 +17,7 @@
 
 - (void).cxx_destruct;
 - (id)_batchOfWorkItems;
-- (void)_processWorkItemsWithDeadline:(double)arg1;
+- (void)_processWorkItemsWithDeadline:(unsigned long long)arg1;
 - (void)_removeWorkItems:(id)arg1;
 - (void)_removeWorkItem:(id)arg1;
 - (BOOL)_workIsEnqueued;
